tests/wpt/web-platform-tests/css/css-break/parsing/break-after-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fragmentation Module Level 3: getComputedValue().breakAfter</title>
<link rel="help" href="https://drafts.csswg.org/css-break/#propdef-break-after">
<meta name="assert" content="break-after computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("break-after", "auto");
test_computed_value("break-after", "avoid");
test_computed_value("break-after", "avoid-page");
test_computed_value("break-after", "page");
test_computed_value("break-after", "left");
test_computed_value("break-after", "right");
test_computed_value("break-after", "recto");
test_computed_value("break-after", "verso");
test_computed_value("break-after", "avoid-column");
test_computed_value("break-after", "column");
test_computed_value("break-after", "avoid-region");
test_computed_value("break-after", "region");
</script>
</body>
</html>
